requirements.txt
beautifulsoup4
requests
newspaper3k
lxml_html_clean
pandas
sentence-transformers